+++ src/Transformers/BM25Transformer.php
@@ -5,9 +5,8 @@ namespace Rubix\ML\Transformers;
 use Rubix\ML\DataType;
 use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Specifications\SamplesAreCompatibleWithTransformer;
-use InvalidArgumentException;
-use RuntimeException;
-use Stringable;
+use Rubix\ML\Exceptions\InvalidArgumentException;
+use Rubix\ML\Exceptions\RuntimeException;
 
 use function is_null;
 
@@ -27,17 +26,17 @@ use function is_null;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class BM25Transformer implements Transformer, Stateful, Elastic, Stringable
+class BM25Transformer implements Transformer, Stateful, Elastic
 {
     /**
-     * The term frequency (TF) normalization factor.
+     * The term frequency (TF) saturation factor. Lower values will cause TF to saturate quicker.
      *
      * @var float
      */
     protected $alpha;
 
     /**
-     * The importance of document length in normalizing term frequency.
+     * The importance of document length in normalizing the term frequency.
      *
      * @var float
      */
@@ -82,7 +81,7 @@ class BM25Transformer implements Transformer, Stateful, Elastic, Stringable
     /**
      * @param float $alpha
      * @param float $beta
-     * @throws \InvalidArgumentException
+     * @throws \Rubix\ML\Exceptions\InvalidArgumentException
      */
     public function __construct(float $alpha = 1.2, float $beta = 0.75)
     {
@@ -160,7 +159,7 @@ class BM25Transformer implements Transformer, Stateful, Elastic, Stringable
      * Update the fitting of the transformer.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
-     * @throws \InvalidArgumentException
+     * @throws \Rubix\ML\Exceptions\InvalidArgumentException
      */
     public function update(Dataset $dataset) : void
     {
@@ -199,7 +198,7 @@ class BM25Transformer implements Transformer, Stateful, Elastic, Stringable
      * Transform the dataset in place.
      *
      * @param array[] $samples
-     * @throws \RuntimeException
+     * @throws \Rubix\ML\Exceptions\RuntimeException
      */
     public function transform(array &$samples) : void
     {