+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - 0.3.0-beta
+    - Added Gzip serializer
     - Move K Best Selector to main repository
 
 - 0.2.1-beta