+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-- 0.2.0-beta1
+- 0.3.0-beta
+    - Added Gzip and Bzip2 serializers
+    - Move K Best Selector to main repository
+
+- 0.2.1-beta
+    - Implemented K Best feature selector
+
+- 0.2.0-beta
     - Implemented Token Hashing Vectorizer transformer
     - Move Recursive Feature Eliminator (RFE) to main repo
     - Added Laplace smoothing to Delta TF-IDF Transformer